#14: adds Disqus comment system
This commit tries to fix #14. It simply adds a Disqus comment system under
the existing Facebook comment system without tuning any css style.
The div box will fill the width of the whole available area.

Please change the Disqus shortname from 'example' to an actually registred
Disqus shortname in templates/base.html.

This commit is in not yet a complete fix, but more like a demonstration of
how to use Disqus in the website.